Fix organization of Beam Jobs (#13771)
