+++ jobs/batch_jobs/validation_jobs_test.py
@@ -23,7 +23,7 @@ from core.platform import models
 import feconf
 from jobs import job_test_utils
 from jobs.batch_jobs import validation_jobs
-from jobs.transforms import base_validation
+from jobs.transforms.validation import base_validation
 from jobs.types import base_validation_errors
 from jobs.types import model_property
