+++ jobs/transforms/validation/base_validation_registry.py
@@ -29,10 +29,20 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 
 from jobs.decorators import validation_decorators
-from jobs.transforms import auth_validation  # pylint: disable=unused-import
-from jobs.transforms import base_validation  # pylint: disable=unused-import
-from jobs.transforms import blog_validation  # pylint: disable=unused-import
-from jobs.transforms import user_validation  # pylint: disable=unused-import
+from jobs.transforms.validation import auth_validation         # pylint: disable=unused-import  # isort: skip
+from jobs.transforms.validation import base_validation         # pylint: disable=unused-import  # isort: skip
+from jobs.transforms.validation import blog_validation         # pylint: disable=unused-import  # isort: skip
+from jobs.transforms.validation import collection_validation   # pylint: disable=unused-import  # isort: skip
+from jobs.transforms.validation import config_validation       # pylint: disable=unused-import  # isort: skip
+from jobs.transforms.validation import exp_validation          # pylint: disable=unused-import  # isort: skip
+from jobs.transforms.validation import feedback_validation     # pylint: disable=unused-import  # isort: skip
+from jobs.transforms.validation import improvements_validation # pylint: disable=unused-import  # isort: skip
+from jobs.transforms.validation import question_validation     # pylint: disable=unused-import  # isort: skip
+from jobs.transforms.validation import skill_validation        # pylint: disable=unused-import  # isort: skip
+from jobs.transforms.validation import story_validation        # pylint: disable=unused-import  # isort: skip
+from jobs.transforms.validation import subtopic_validation     # pylint: disable=unused-import  # isort: skip
+from jobs.transforms.validation import topic_validation        # pylint: disable=unused-import  # isort: skip
+from jobs.transforms.validation import user_validation         # pylint: disable=unused-import  # isort: skip
 
 
 def get_audit_do_fn_types_by_kind():
