+++ jobs/transforms/validation/base_validation_registry_test.py
@@ -21,7 +21,7 @@ from __future__ import unicode_literals
 
 from core.tests import test_utils
 from jobs.decorators import validation_decorators
-from jobs.transforms import base_validation_registry
+from jobs.transforms.validation import base_validation_registry
 
 
 class GetAuditsByKindTests(test_utils.TestBase):
