+++ jobs/transforms/validation/blog_validation.py
@@ -23,10 +23,10 @@ import datetime
 
 from core.domain import blog_domain
 from core.platform import models
-from jobs import blog_validation_errors
 from jobs import job_utils
 from jobs.decorators import validation_decorators
-from jobs.transforms import base_validation
+from jobs.transforms.validation import base_validation
+from jobs.types import blog_validation_errors
 
 import apache_beam as beam
