+++ jobs/transforms/validation/config_validation.py
@@ -23,7 +23,7 @@ from core.domain import config_domain
 from core.domain import platform_parameter_domain as parameter_domain
 from core.platform import models
 from jobs.decorators import validation_decorators
-from jobs.transforms import base_validation
+from jobs.transforms.validation import base_validation
 
 (config_models,) = models.Registry.import_models([models.NAMES.config])
