+++ jobs/transforms/validation/config_validation_test.py
@@ -22,7 +22,7 @@ from __future__ import unicode_literals
 from core.domain import platform_parameter_domain as parameter_domain
 from core.platform import models
 from jobs import job_test_utils
-from jobs.transforms import config_validation
+from jobs.transforms.validation import config_validation
 from jobs.types import base_validation_errors
 
 import apache_beam as beam
