+++ jobs/transforms/validation/exp_validation_test.py
@@ -24,7 +24,7 @@ from core.platform import models
 from core.tests import test_utils
 from jobs import job_test_utils
 from jobs.decorators import validation_decorators
-from jobs.transforms import exp_validation
+from jobs.transforms.validation import exp_validation
 from jobs.types import base_validation_errors
 
 import apache_beam as beam
