+++ jobs/transforms/validation/improvements_validation_test.py
@@ -23,7 +23,7 @@ from __future__ import unicode_literals
 from core.platform import models
 
 from jobs import job_test_utils
-from jobs.transforms import improvements_validation
+from jobs.transforms.validation import improvements_validation
 from jobs.types import improvements_validation_errors
 
 import apache_beam as beam
