+++ jobs/transforms/validation/subtopic_validation_test.py
@@ -21,7 +21,7 @@ from __future__ import unicode_literals
 
 from core.platform import models
 from jobs import job_test_utils
-from jobs.transforms import subtopic_validation
+from jobs.transforms.validation import subtopic_validation
 from jobs.types import base_validation_errors
 
 import apache_beam as beam
