+++ jobs/transforms/validation/topic_validation_test.py
@@ -26,7 +26,7 @@ from core.platform import models
 from core.tests import test_utils
 from jobs import job_test_utils
 from jobs.decorators import validation_decorators
-from jobs.transforms import topic_validation
+from jobs.transforms.validation import topic_validation
 from jobs.types import base_validation_errors
 from jobs.types import topic_validation_errors
