+++ jobs/transforms/validation/user_validation.py
@@ -25,7 +25,7 @@ from core.platform import models
 import feconf
 from jobs import job_utils
 from jobs.decorators import validation_decorators
-from jobs.transforms import base_validation
+from jobs.transforms.validation import base_validation
 from jobs.types import user_validation_errors
 
 import apache_beam as beam
