+++ jobs/transforms/validation/user_validation_test.py
@@ -26,7 +26,7 @@ from core.tests import test_utils
 import feconf
 from jobs import job_test_utils
 from jobs.decorators import validation_decorators
-from jobs.transforms import user_validation
+from jobs.transforms.validation import user_validation
 from jobs.types import base_validation_errors
 from jobs.types import user_validation_errors
