+++ scripts/backend_test_shards.json
@@ -153,11 +153,11 @@
         "scripts.release_scripts.cut_release_or_hotfix_branch_test",
         "scripts.linters.css_linter_test",
         "core.platform.email.dev_mode_email_services_test",
-        "jobs.blog_validation_errors_test",
-        "jobs.blog_validation_jobs_test",
-        "jobs.transforms.blog_validation_test",
-        "jobs.transforms.skill_validation_test",
-        "jobs.transforms.story_validation_test",
+        "jobs.types.blog_validation_errors_test",
+        "jobs.batch_jobs.blog_validation_jobs_test",
+        "jobs.transforms.validation.blog_validation_test",
+        "jobs.transforms.validation.skill_validation_test",
+        "jobs.transforms.validation.story_validation_test",
         "core.tests.load_tests.feedback_thread_summaries_test",
         "core.tests.test_utils_test",
         "core.tests.gae_suite_test"
@@ -243,18 +243,18 @@
         "jobs.job_test_utils_test",
         "jobs.job_utils_test",
         "jobs.registry_test",
-        "jobs.transforms.base_validation_registry_test",
-        "jobs.transforms.exp_validation_test",
-        "jobs.transforms.topic_validation_test",
-        "jobs.transforms.subtopic_validation_test",
-        "jobs.transforms.auth_validation_test",
-        "jobs.transforms.base_validation_test",
-        "jobs.transforms.user_validation_test",
-        "jobs.transforms.config_validation_test",
-        "jobs.transforms.collection_validation_test",
-        "jobs.transforms.question_validation_test",
-        "jobs.transforms.feedback_validation_test",
-        "jobs.transforms.improvements_validation_test",
+        "jobs.transforms.validation.base_validation_registry_test",
+        "jobs.transforms.validation.exp_validation_test",
+        "jobs.transforms.validation.topic_validation_test",
+        "jobs.transforms.validation.subtopic_validation_test",
+        "jobs.transforms.validation.auth_validation_test",
+        "jobs.transforms.validation.base_validation_test",
+        "jobs.transforms.validation.user_validation_test",
+        "jobs.transforms.validation.config_validation_test",
+        "jobs.transforms.validation.collection_validation_test",
+        "jobs.transforms.validation.question_validation_test",
+        "jobs.transforms.validation.feedback_validation_test",
+        "jobs.transforms.validation.improvements_validation_test",
         "jobs.types.base_validation_errors_test",
         "jobs.types.job_run_result_test",
         "jobs.types.model_property_test",
