+++ scripts/run_mypy_checks.py
@@ -506,10 +506,10 @@ NOT_FULLY_COVERED_FILES = [
     'jobs/base_jobs_test.py',
     'jobs/batch_jobs/validation_jobs.py',
     'jobs/batch_jobs/validation_jobs_test.py',
-    'jobs/blog_validation_errors.py',
-    'jobs/blog_validation_errors_test.py',
-    'jobs/blog_validation_jobs.py',
-    'jobs/blog_validation_jobs_test.py',
+    'jobs/types/blog_validation_errors.py',
+    'jobs/types/blog_validation_errors_test.py',
+    'jobs/batch_jobs/blog_validation_jobs.py',
+    'jobs/batch_jobs/blog_validation_jobs_test.py',
     'jobs/decorators/validation_decorators.py',
     'jobs/decorators/validation_decorators_test.py',
     'jobs/io/job_io.py',
@@ -524,36 +524,36 @@ NOT_FULLY_COVERED_FILES = [
     'jobs/job_utils_test.py',
     'jobs/registry.py',
     'jobs/registry_test.py',
-    'jobs/transforms/auth_validation.py',
-    'jobs/transforms/auth_validation_test.py',
-    'jobs/transforms/base_validation.py',
-    'jobs/transforms/base_validation_registry.py',
-    'jobs/transforms/base_validation_registry_test.py',
-    'jobs/transforms/base_validation_test.py',
-    'jobs/transforms/blog_validation.py',
-    'jobs/transforms/blog_validation_test.py',
-    'jobs/transforms/collection_validation.py',
-    'jobs/transforms/collection_validation_test.py',
-    'jobs/transforms/config_validation.py',
-    'jobs/transforms/config_validation_test.py',
-    'jobs/transforms/exp_validation.py',
-    'jobs/transforms/exp_validation_test.py',
-    'jobs/transforms/feedback_validation.py',
-    'jobs/transforms/feedback_validation_test.py',
-    'jobs/transforms/improvements_validation.py',
-    'jobs/transforms/improvements_validation_test.py',
-    'jobs/transforms/question_validation.py',
-    'jobs/transforms/question_validation_test.py',
-    'jobs/transforms/skill_validation.py',
-    'jobs/transforms/skill_validation_test.py',
-    'jobs/transforms/story_validation.py',
-    'jobs/transforms/story_validation_test.py',
-    'jobs/transforms/subtopic_validation.py',
-    'jobs/transforms/subtopic_validation_test.py',
-    'jobs/transforms/topic_validation.py',
-    'jobs/transforms/topic_validation_test.py',
-    'jobs/transforms/user_validation.py',
-    'jobs/transforms/user_validation_test.py',
+    'jobs/transforms/validation/auth_validation.py',
+    'jobs/transforms/validation/auth_validation_test.py',
+    'jobs/transforms/validation/base_validation.py',
+    'jobs/transforms/validation/base_validation_registry.py',
+    'jobs/transforms/validation/base_validation_registry_test.py',
+    'jobs/transforms/validation/base_validation_test.py',
+    'jobs/transforms/validation/blog_validation.py',
+    'jobs/transforms/validation/blog_validation_test.py',
+    'jobs/transforms/validation/collection_validation.py',
+    'jobs/transforms/validation/collection_validation_test.py',
+    'jobs/transforms/validation/config_validation.py',
+    'jobs/transforms/validation/config_validation_test.py',
+    'jobs/transforms/validation/exp_validation.py',
+    'jobs/transforms/validation/exp_validation_test.py',
+    'jobs/transforms/validation/feedback_validation.py',
+    'jobs/transforms/validation/feedback_validation_test.py',
+    'jobs/transforms/validation/improvements_validation.py',
+    'jobs/transforms/validation/improvements_validation_test.py',
+    'jobs/transforms/validation/question_validation.py',
+    'jobs/transforms/validation/question_validation_test.py',
+    'jobs/transforms/validation/skill_validation.py',
+    'jobs/transforms/validation/skill_validation_test.py',
+    'jobs/transforms/validation/story_validation.py',
+    'jobs/transforms/validation/story_validation_test.py',
+    'jobs/transforms/validation/subtopic_validation.py',
+    'jobs/transforms/validation/subtopic_validation_test.py',
+    'jobs/transforms/validation/topic_validation.py',
+    'jobs/transforms/validation/topic_validation_test.py',
+    'jobs/transforms/validation/user_validation.py',
+    'jobs/transforms/validation/user_validation_test.py',
     'jobs/types/base_validation_errors.py',
     'jobs/types/base_validation_errors_test.py',
     'jobs/types/feedback_validation_errors.py',
